Upgrade to Discord.js v14, add PulseAudio mode

+++ src/stop.ts
@@ -1,10 +1,24 @@
 import { client } from './discord-client';
 
-export function stop(graceful: boolean): never {
+let stopping = false;
+
+export function stop(graceful: boolean): void {
+    if (stopping) {
+        process.exit(1);
+    }
+    stopping = true;
+
     try {
-        client.destroy();
-    } catch (e) { }
-    process.exit(graceful ? 0 : 1);
+        client.destroy()
+            .then(() => {
+                process.exit(graceful ? 0 : 1);
+            })
+            .catch(() => {
+                process.exit(1);
+            });
+    } catch (e) {
+        process.exit(1);
+    }
 }
 
 process.on('uncaughtException', error => {
@@ -17,5 +31,11 @@ process.on('unhandledRejection', reason => {
 });
 
 process.on('SIGINT', () => {
+    stop(true);
+});
+process.on('SIGTERM', () => {
+    stop(true);
+});
+process.on('SIGHUP', () => {
     stop(true);
 });
